requirements.txt
exceptiongroup==1.3.1
grpcio==1.80.0
grpcio-tools==1.80.0
iniconfig==2.1.0
packaging==26.1
pluggy==1.6.0
protobuf==6.33.6
Pygments==2.20.0
pytest==8.4.2
tomli==2.4.1
typing_extensions==4.15.0
